% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxaIDs.R
\name{getTaxaIDs}
\alias{getTaxaIDs}
\title{getTaxaIDs}
\usage{
getTaxaIDs(spec_list = NULL, sci_col = NULL, comm_col = NULL)
}
\arguments{
\item{spec_list}{the dataframe containing information to be decoded to TSN and
aphiaIDs}

\item{sci_col}{the name of the column of the dataframe containing the
scientific names}

\item{comm_col}{the name of the column of the dataframe containing the
common names}
}
\description{
This function attempts to determine the aphiaIDs (WoRMS) and
TSNs (ITIS) using first the scientific name, and then the common name. If an
aphiaid is found, but no TSN, it will try to find the TSN using the aphiaid.
If a TSN is found, but no aphiaid, it will try to find the aphiaid using the
TSN.
Additionally, in addition to all of the original fields, the returned data
will also include the WoRMS aphiaid, the ITIS TSN, and will indicate the
method used to find each.  Also for each, a field will indicate whether one,
many, or no matches were found.  Lastly, if the species has a different
"accepted name" than the one provided, that will also be returned.
}
\seealso{
Other qc: \code{\link{qc_fieldvalues}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
