% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_fieldvalues.r
\name{qc_fieldvalues}
\alias{qc_fieldvalues}
\title{qc_fieldvalues}
\usage{
qc_fieldvalues(qc.df = NULL, field.master = NULL,
  field.master.values = NULL, fields.qc = NULL, min.n = 5)
}
\arguments{
\item{qc.df}{default is \code{NULL}. This identifies the name of the table
you want to plot values from.}

\item{field.master}{default is \code{NULL}.  This identifies the field you
want to generates plots by.  In the example above, this would be SPECIES.
param field.master.values default is \code{NULL}.  In cases where you only
want to QC a few different potential values for\code{field.master}, you can
specify them here.}

\item{field.master.values}{default is \code{NULL}. If you are only interested
in a few values of your \code{field.master}, you can enter them as a vector
here (e.g. a couple years, or a list of species codes, etc)}

\item{fields.qc}{default is \code{NULL}. This is a vector of between 2 and 5
fields you want to plot against eachother.  If left blank, and there are less
than 6 fields, they will all be plotted.}

\item{min.n}{default is 5.  This is the minimum number of values you need
before plots will be done}
}
\description{
This function takes a table, and for every unque value of one
field, plots the values of other specified fields against eachother.  For
example, you might QC you catch table by species - in which case, for every
different species, you might plot the TOTNO vs the TOTWGT.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
