% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_findorphans.r
\name{qc_findorphans}
\alias{qc_findorphans}
\title{qc_findorphans}
\usage{
qc_findorphans(db = NULL)
}
\arguments{
\item{db}{default is \code{NULL}. This identifies the dataset you are working
with.}
}
\description{
This function creates orph_* tables with records that don't
appear to be linked to the other tables (within bio.datawrangling)
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
