% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outlierCheck.R
\name{qc_outlierCheck}
\alias{qc_outlierCheck}
\title{qc_outlierCheck}
\usage{
qc_outlierCheck(df, field)
}
\arguments{
\item{df}{a dataframe to be analyzed.}

\item{field}{a fieldname within the df to be analyzed.}
}
\value{
the df, with an additional field - "OL_<field>".
}
\description{
This function takes a dataframe, and does a simple check for 
outliers based on the boxplot.stats() routine. Specifically, it identifies 
data which is 1.5, 5, or 10 times the length of the box away from the box.  
These are identified as "Possible", "Likely", or "Extreme" in the resultant 
"OL_<field>" column, respectively. Other data is identified as "OK", which 
indicates that it is within a reasonable range of the mean.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
